@@ -1,6 +1,6 @@
 module simple-proxy
 
-go 1.23.2
+go 1.24.1
 
 replace go.minekube.com/gate => ../../../
 
