@@ -1,11 +1,11 @@
 module example
 
-go 1.23.2
+go 1.24.1
 
 require (
 	buf.build/gen/go/minekube/gate/connectrpc/go v1.18.1-20241120101512-f1a10b5029ce.1
-	buf.build/gen/go/minekube/gate/protocolbuffers/go v1.36.4-20241120101512-f1a10b5029ce.1
+	buf.build/gen/go/minekube/gate/protocolbuffers/go v1.36.6-20241120101512-f1a10b5029ce.1
 	connectrpc.com/connect v1.18.1
 )
 
-require google.golang.org/protobuf v1.36.4 // indirect
+require google.golang.org/protobuf v1.36.6 // indirect
