@@ -1,34 +1,32 @@
 module go.minekube.com/gate
 
-go 1.23.2
-
-toolchain go1.23.5
+go 1.24.1
 
 require (
 	connectrpc.com/connect v1.18.1
 	github.com/Tnze/go-mc v1.20.2
 	github.com/agext/levenshtein v1.2.3
-	github.com/coder/websocket v1.8.12
+	github.com/coder/websocket v1.8.13
 	github.com/dboslee/lru v0.0.1
 	github.com/edwingeng/deque/v2 v2.1.1
 	github.com/gammazero/deque v1.0.0
-	github.com/go-faker/faker/v4 v4.5.0
+	github.com/go-faker/faker/v4 v4.6.1
 	github.com/go-logr/logr v1.4.2
 	github.com/go-logr/zapr v1.3.0
-	github.com/golang/groupcache v0.0.0-20210331224755-41bb18bfe9da
+	github.com/golang/groupcache v0.0.0-20241129210726-2c02b8208cf8
 	github.com/google/uuid v1.6.0
 	github.com/gookit/color v1.5.4
 	github.com/jellydator/ttlcache/v3 v3.3.0
-	github.com/knadh/koanf/providers/file v1.1.2
+	github.com/knadh/koanf/providers/file v1.2.0
 	github.com/nfnt/resize v0.0.0-20180221191011-83c6a9932646
 	github.com/pires/go-proxyproto v0.8.0
 	github.com/robinbraemer/event v0.1.1
 	github.com/rs/xid v1.6.0
 	github.com/sandertv/go-raknet v1.13.0
 	github.com/sandertv/gophertunnel v1.37.0
-	github.com/spf13/viper v1.19.0
+	github.com/spf13/viper v1.20.1
 	github.com/stretchr/testify v1.10.0
-	github.com/urfave/cli/v2 v2.27.5
+	github.com/urfave/cli/v2 v2.27.6
 	github.com/zyedidia/generic v1.2.1
 	go.minekube.com/brigodier v0.0.1
 	go.minekube.com/common v0.0.6
