@@ -47,6 +47,7 @@ var (
 	Minecraft_1_21   = v(767, "1.21", "1.21.1")
 	Minecraft_1_21_2 = v(768, "1.21.2", "1.21.3")
 	Minecraft_1_21_4 = v(769, "1.21.4")
+	Minecraft_1_21_5 = v(770, "1.21.5")
 
 	// Versions ordered from lowest to highest
 	Versions = []*proto.Version{
@@ -66,7 +67,7 @@ var (
 		Minecraft_1_18, Minecraft_1_18_2,
 		Minecraft_1_19, Minecraft_1_19_1, Minecraft_1_19_3, Minecraft_1_19_4,
 		Minecraft_1_20, Minecraft_1_20_2, Minecraft_1_20_3, Minecraft_1_20_5,
-		Minecraft_1_21, Minecraft_1_21_2, Minecraft_1_21_4,
+		Minecraft_1_21, Minecraft_1_21_2, Minecraft_1_21_4, Minecraft_1_21_5,
 	}
 )
 
